requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
sqlalchemy
pika
python-dotenv
ffmpeg-python
ldap3==2.9.1
PySide6==6.7.0
cryptography==42.0.5
requests==2.31.0